third_party/WebKit/LayoutTests/inspector/timeline/timeline-gc-event-expected.txt

Tests the Timeline API instrumentation of a gc event

The expected GC event record
{
    counters : <object>
    data : {
    }
    endTime : <number>
    startTime : <number>
    type : "GCEvent"
}
